move the quinn flume socket into a separate optional module
