+++ tests/interprocess.rs
@@ -84,9 +84,12 @@ pub fn make_endpoints() -> anyhow::Result<Endpoints> {
     let client_addr: SocketAddr = SocketAddr::V4(SocketAddrV4::new(Ipv4Addr::LOCALHOST, 2));
     let (server_config, server_cert) = configure_server()?;
     let client_config = configure_client(&[&server_cert])?;
-    let (server, mut client) =
-        quic_rpc::transport::interprocess::endpoint_pair(server_addr, client_addr, server_config)
-            .unwrap();
+    let (server, mut client) = quic_rpc::transport::quinn_flume_socket::endpoint_pair(
+        server_addr,
+        client_addr,
+        server_config,
+    )
+    .unwrap();
     client.set_default_client_config(client_config);
     Ok(Endpoints {
         client,
